test(terminal): split 948-line test_repositories.py by repository class (#113)
`services/terminal/tests/unit/test_repositories.py` had grown to 948
lines with 5 distinct repository test classes glued together.

Split by repository:
- test_terminal_info_repository.py        (217 lines)
- test_tenant_info_repository.py          (352 lines)
- test_cash_in_out_log_repository.py       (99 lines)
- test_open_close_log_repository.py        (86 lines)
- test_terminallog_delivery_status_repository.py (218 lines)

Module-level helpers (_mock_db, _terminal_info_doc, _tenant_info_doc,
_store_info, _cash_log, _open_close_log) moved to a co-located
`_helpers.py` (no `test_` prefix so pytest does not collect it). Each
new file imports only the helpers it actually uses.

No test logic changed. Pass count unchanged: terminal unit tier
332/332 passes (was 332/332 before split).

Refs #113
